refactor: Improve timeout handling and browser compatibility
- Remove specification of 0 timeout behavior, default to no timeout
- Replace NodeJS.Timeout with browser-compatible type
- Move timeout utility to ai package under util directory
- Ensure timeout clearing occurs in finally block for error handling
