+++ content/docs/07-reference/ai-sdk-core/01-generate-text.mdx
@@ -337,7 +337,7 @@ To see `generateText` in action, check out [these examples](#examples).
       type: 'number',
       isOptional: true,
       description:
-        'Timeout in milliseconds. If set, the call will be aborted after the timeout. Defaults to 0 (no timeout).',
+        'Timeout in milliseconds. If set, the call will be aborted after the specified duration. By default, there is no timeout.',
     },
     {
       name: 'abortSignal',
