+++ content/docs/07-reference/ai-sdk-core/03-generate-object.mdx
@@ -401,7 +401,7 @@ To see `generateObject` in action, check out the [additional examples](#more-exa
       type: 'number',
       isOptional: true,
       description:
-        'Timeout in milliseconds. If set, the call will be aborted after the timeout. Defaults to 0 (no timeout).',
+        'Timeout in milliseconds. If set, the call will be aborted after the specified duration. By default, there is no timeout.',
     },
     {
       name: 'abortSignal',
