+++ packages/provider-utils/src/index.ts
@@ -15,4 +15,3 @@ export * from './uint8-utils';
 export * from './validate-types';
 export * from './validator';
 export * from './without-trailing-slash';
-export * from './get-effective-abort-signal';
